Add TestingExtendedHiveMetastore
